Refactor - Move Born of the Gods tokens to shared area
