+++ Mage.Sets/src/mage/sets/bornofthegods/PheresBandRaiders.java
@@ -36,7 +36,7 @@ import mage.abilities.keyword.InspiredAbility;
 import mage.cards.CardImpl;
 import mage.constants.CardType;
 import mage.constants.Rarity;
-import mage.sets.bornofthegods.TokenAndCounters.CentaurEnchantmentCreatureToken;
+import mage.game.permanent.token.CentaurEnchantmentCreatureToken;
 
 /**
  *
