+++ Mage.Sets/src/mage/sets/journeyintonyx/KingMacarTheGoldCursed.java
@@ -37,7 +37,7 @@ import mage.abilities.keyword.InspiredAbility;
 import mage.cards.CardImpl;
 import mage.constants.CardType;
 import mage.constants.Rarity;
-import mage.sets.bornofthegods.TokenAndCounters.GoldToken;
+import mage.game.permanent.token.GoldToken;
 import mage.target.common.TargetCreaturePermanent;
