+++ Mage/src/main/java/mage/game/permanent/token/CentaurEnchantmentCreatureToken.java
@@ -25,10 +25,9 @@
  *  authors and should not be interpreted as representing official policies, either expressed
  *  or implied, of BetaSteward_at_googlemail.com.
  */
-package mage.sets.bornofthegods.TokenAndCounters;
+package mage.game.permanent.token;
 
 import mage.MageInt;
-import mage.abilities.keyword.VigilanceAbility;
 import mage.constants.CardType;
 import mage.game.permanent.token.Token;
 
@@ -36,19 +35,16 @@ import mage.game.permanent.token.Token;
  *
  * @author LevelX2
  */
-public class CatSoldierCreatureToken extends Token {
+public class CentaurEnchantmentCreatureToken extends Token {
 
-    public CatSoldierCreatureToken() {
-        super("Cat Soldier", "1/1 white Cat Soldier creature token with vigilance");
+    public CentaurEnchantmentCreatureToken() {
+        super("Centaur", "3/3 green Centaur enchantment creature token");
+        cardType.add(CardType.ENCHANTMENT);
         cardType.add(CardType.CREATURE);
-        color.setWhite(true);
-        
-        subtype.add("Cat");
-        subtype.add("Soldier");
-        power = new MageInt(1);
-        toughness = new MageInt(1);
-        
-        this.addAbility(VigilanceAbility.getInstance());
+        color.setGreen(true);
+        subtype.add("Centaur");
+        power = new MageInt(3);
+        toughness = new MageInt(3);
         this.setOriginalExpansionSetCode("BNG");
     }
 }
